+++ lib/karafka/connection/messages_buffer.rb
@@ -2,19 +2,26 @@
 
 module Karafka
   module Connection
-    # Buffer for messages.
-    # When message is added to this buffer, it gets assigned to an array with other messages from
-    # the same topic and partition.
+    # Buffer used to build and store karafka messages built based on raw librdkafka messages.
     #
-    # @note This buffer is NOT threadsafe.
+    # Why do we have two buffers? `RawMessagesBuffer` is used to store raw messages and to handle
+    #   cases related to partition revocation and reconnections. It is "internal" to the listening
+    #   process. `MessagesBuffer` on the other hand is used to "translate" those raw messages that
+    #   we know that are ok into Karafka messages and to simplify further work with them.
     #
-    # @note We store data here in groups per topic partition to handle the revocation case, where
-    #    we may need to remove messages from a single topic partition
+    # While it adds a bit of overhead, it makes conceptual things much easier and it adds only two
+    #   simple hash iterations over messages batch.
+    #
+    # @note This buffer is NOT thread safe. We do not worry about it as we do not use it outside
+    #   of the main listener loop. It can be cleared after the jobs are scheduled with messages
+    #   it stores, because messages arrays are not "cleared" in any way directly and their
+    #   reference stays.
     class MessagesBuffer
       attr_reader :size
 
-      # @return [Karafka::Connection::MessagesBuffer] buffer instance
-      def initialize
+      # @param subscription_group [Karafka::Routing::SubscriptionGroup]
+      def initialize(subscription_group)
+        @subscription_group = subscription_group
         @size = 0
         @groups = Hash.new do |topic_groups, topic|
           topic_groups[topic] = Hash.new do |partition_groups, partition|
@@ -23,20 +30,35 @@ module Karafka
         end
       end
 
-      # Adds a message to the buffer.
-      #
-      # @param message [Rdkafka::Consumer::Message] raw rdkafka message
-      # @return [Array<Rdkafka::Consumer::Message>] given partition topic sub-buffer array
-      def <<(message)
-        @size += 1
-        @groups[message.topic][message.partition] << message
+      # Remaps raw messages from the raw messages buffer to Karafka messages
+      # @param raw_messages_buffer [RawMessagesBuffer] buffer with raw messages
+      def remap(raw_messages_buffer)
+        clear unless @size.zero?
+
+        # Since it happens "right after" we've received the messages, it is close enough it time
+        # to be used as the moment we received messages.
+        received_at = Time.now
+
+        raw_messages_buffer.each do |topic, partition, messages|
+          @size += messages.count
+
+          ktopic = @subscription_group.topics.find(topic)
+
+          @groups[topic][partition] = messages.map do |message|
+            Messages::Builders::Message.call(
+              message,
+              ktopic,
+              received_at
+            )
+          end
+        end
       end
 
       # Allows to iterate over all the topics and partitions messages
       #
       # @yieldparam [String] topic name
       # @yieldparam [Integer] partition number
-      # @yieldparam [Array<Rdkafka::Consumer::Message>] topic partition aggregated results
+      # @yieldparam [Array<Karafka::Messages::Message>] messages from a given topic partition
       def each
         @groups.each do |topic, partitions|
           partitions.each do |partition, messages|
@@ -45,55 +67,17 @@ module Karafka
         end
       end
 
-      # Removes given topic and partition data out of the buffer
-      # This is used when there's a partition revocation
-      # @param topic [String] topic we're interested in
-      # @param partition [Integer] partition of which data we want to remove
-      def delete(topic, partition)
-        return unless @groups.key?(topic)
-        return unless @groups.fetch(topic).key?(partition)
-
-        topic_data = @groups.fetch(topic)
-        topic_data.delete(partition)
-
-        recount!
-
-        # If there are no more partitions to handle in a given topic, remove it completely
-        @groups.delete(topic) if topic_data.empty?
-      end
-
-      # Removes duplicated messages from the same partitions
-      # This should be used only when rebalance occurs, as we may get data again we already have
-      # due to the processing from the last offset. In cases like this, we may get same data
-      # again and we do want to ensure as few duplications as possible
-      def uniq!
-        @groups.each_value do |partitions|
-          partitions.each_value do |messages|
-            messages.uniq!(&:offset)
-          end
-        end
-
-        recount!
-      end
-
-      # Removes all the data from the buffer.
-      #
-      # @note We do not clear the whole groups hash but rather we clear the partition hashes, so
-      #   we save ourselves some objects allocations. We cannot clear the underlying arrays as they
-      #   may be used in other threads for data processing, thus if we would clear it, we could
-      #   potentially clear a raw messages array for a job that is in the jobs queue.
-      def clear
-        @size = 0
-        @groups.each_value(&:clear)
+      # @return [Boolean] is the buffer empty or does it contain any messages
+      def empty?
+        @size.zero?
       end
 
       private
 
-      # Updates the messages count if we performed any operations that could change the state
-      def recount!
-        @size = @groups.each_value.sum do |partitions|
-          partitions.each_value.map(&:count).sum
-        end
+      # Clears the buffer completely
+      def clear
+        @size = 0
+        @groups.clear
       end
     end
   end